pyproject.toml
[project]
name = "simple_bert"
description = "A package for training small and simple bert models"
readme = "README.md"
authors = [{ name = "Sergio Frayle Perez", email = "sfp932705@gmail.com" }]
version = "0.0.8"
requires-python = ">=3.13"
dependencies = ["numpy==2.4.1", "torch==2.9.1", "pydantic-settings==2.12.0"]

[project.optional-dependencies]
test = ["mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-dotenv>=0.5.2",
    "pytest-mock>=3.15.1",
    "ruff>=0.14.10",
]

[tool.black]
line-length = 88

[tool.bumpversion]
commit = true
tag = false

[tool.mypy]
mypy_path = ["src"]
ignore_missing_imports = false
check_untyped_defs = true
exclude = '^data\/*'
explicit_package_bases = true

[tool.pytest.ini_options]
addopts = "-p no:cacheprovider --cov-report term-missing"
env_files = ".env.test"
testpaths = ["tests"]

[tool.coverage.report]
skip_empty = true
skip_covered = true
omit = ["src/main.py"]
exclude_also = [
    "def __repr__",
    "@(abc\\.)?abstractmethod",
    "__name__ == .__main__.",
    "if TYPE_CHECKING:",
    "class .*\bException\\):",
    "class .*\bError\\):",
    "class .*\\(abc\\.ABC\\):",
]


[[tool.mypy.overrides]]
module = ["rust.*"]
ignore_missing_imports = true

[tool.ruff]
exclude = [".git", "__pycache__", ".venv", "venv", ]
line-length = 88
indent-width = 4
src = ["src/", "tests/"]

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "E501", "I001"]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[build-system]
requires = ["maturin>=1.4,<2.0"]
build-backend = "maturin"

[tool.maturin]
python-source = "src"
module-name = "simple_bert.rust"
manifest-path = "rust/Cargo.toml"
